Complete data insert script, update schema
Updated the schema to allow for greater precision because open and close
prices demanded it. The data insertion script will now insert for any
valid series of stock tickers and index tickers. It does take a few
minutes, so run it and let it do its thing.

Requires pg8000, yahoo_finance and the built-in getpass (I think it's
built in). It prompts you for your user and pass.